refactor >> context통합, hook함수 util로 분리, priceComma추가했습니다
