+++ src/components/my-page-modal.js
@@ -5,9 +5,10 @@ import AuthApi from "apis/auth";
 import { PRODUCT_QUERY_KEY } from "consts";
 import MiniUserInfo from "./miniuser-Info";
 import { useNavigate } from "react-router-dom";
-import { useAuth } from "provider/auth-provider";
+
 import MMMAlert from "./mmm-alert";
 import { useState } from "react";
+import { useAuth } from "context/auth.ctx";
 
 const MyPageModal = ({ setIsMyPageModal }) => {
   // alert
