+++ src/pages/login-page/components/signIn-form.js
@@ -6,12 +6,10 @@ import MMMInput from "components/input";
 import MMMAlert from "components/mmm-alert";
 import styled from "styled-components";
 import { flexCenter } from "styles/common.style";
-import { useAuth } from "provider/auth-provider";
-import { useSocket } from "socket/socket";
 import LoginUserNickNameRepository from "repository/login-user-nickName-repository";
-import TokenRepository from "repository/token-repository";
 import { SocketTokenRepository } from "repository/socket-token-repository";
 import { useState } from "react";
+import { useAuth } from "context/auth.ctx";
 
 const SignInForm = () => {
   // alert
