+++ src/pages/login-page/components/signUp-form.js
@@ -3,7 +3,6 @@ import { useNavigate } from "react-router-dom";
 import { useRecoilState } from "recoil";
 import { Api } from "apis";
 import AuthApi from "apis/auth";
-import { useAuth } from "provider/auth-provider";
 import { isEmailCheckPass, isNickNameCheckPass } from "store";
 import { FormValidate } from "utils/validate-helper";
 import MMMButton from "components/button";
@@ -15,6 +14,7 @@ import { flexCenter } from "styles/common.style";
 import ProductOrder from "./location2";
 import { useState } from "react";
 import MMMAlert from "components/mmm-alert";
+import { useAuth } from "context/auth.ctx";
 
 const SignUpForm = () => {
   // alert
